Calculates the 95% confidence interval using bootstrapping

Internal function. Not intended for users. This function can return the 95% confidence intervals for any statistic so long as the function passed to f returns a single value of that statistic.

Usage

boot_conf(df, bootstrap_iterations, f, ...)

Arguments

df

a data.frame containing the columns phs, age, and status

bootstrap_iterations

number of bootstrap iterations to run.

f

the name of a function to perform bootstrapping on.

...

arguments to be passed on to f().

Value

A list of confidence intervals