Returns a hazard ratio and 95% CI for Polygenic Hazard Scores

This function calculates the hazard ratio for a 1 standard deviation increase in PHS and optionally performs bootstrap resampling to return 95% confidence intervals. The data can either be provided as a data.frame with columns containing the phs, age, and status of each individual or separate vectors containing each of these values. The columns in data default to 'phs', 'age', and 'status', but any arbitrary column names can be used if named explicitly.

Usage

get_hrsd(
  data = NULL,
  phs = "phs",
  age = "age",
  status = "status",
  conf.int = FALSE,
  conf.level = 0.95,
  bootstrap.iterations
)

Arguments

data

an optional data.frame containing the variables for phs, age, and status

phs

an optional string specifying the column name in data containing the polygenic hazard score for each subject . The default is "phs"

age

an optional string specifying the column name in data containing the age of each subject. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"

status

an optional string specifying the column name in data containing case-control status (0 = censored, 1 = event). The default is "status"

conf.int

logical. if TRUE performs bootstrap and returns 95% confidence intervals. Default = FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

bootstrap.iterations

Number of bootstrap iterations to run. Required if boot = TRUE. Default = 1000.

Value

A numeric hazard ratio or a list containing HR and the 95% confidence intervals from bootstrap

Examples

HR80_20 <- get_hr(test_data, conf.int = TRUE, bootstrap.iterations = 300)