Plot Histogram of Cases and Controls by PHS

This function takes PHS scores and case/control status and plots a histogram.

Usage

phs_hist(
  data = NULL,
  phs = "phs",
  status = "status",
  normalize = FALSE,
  scale = FALSE,
  inverse = FALSE
)

Arguments

data

an optional data.frame containing the variables for phs, age, and status

phs

an optional string specifying the column name in data containing the polygenic hazard score for each subject or the unquoted name of a vector containing these values. The default is "phs"

status

an optional string specifying the column name in data containing case-control status (0 = censored, 1 = event) or the unquoted name of a vector containing these values. The default is "status"

normalize

Boolean indicating whether to normalize the histogram to area = 1. This replaces the y axis with denisty instead of count. Helpful to compare distributions with very different counts. Default = FALSE.

scale

Boolean indicating whether to center and scale the PHS scores to unit variance. Default = FALSE.

inverse

Boolean indicating whether to inverse (x * -1) the PHS scores to reverse the direction of effect. Default = FALSE.

Value

A ggplot object

Examples

phs_hist <- phs_hist(test_data, normalize = TRUE)