Return Kaplan-Meier curve

This function returns a single Kaplan-Meier curve for plotting. Output includes age, the K-M estimate at each age, and the upper and lower confidence intervals

Usage

km_curve(
  data = NULL,
  phs = "phs",
  age = "age",
  status = "status",
  interval = c(0, 1),
  age_range = 40:100,
  scale = FALSE,
  inverse = FALSE
)

Arguments

data

an optional data.frame containing the variables for phs, age, and status

phs

an optional string specifying the column name in data containing the polygenic hazard score for each subject or the unquoted name of a vector containing these values. The default is "phs"

age

an optional string specifying the column name in data containing the age of each subject or the unquoted name of a vector containing these values. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"

status

an optional string specifying the column name in data containing case-control status (0 = censored, 1 = event) or the unquoted name of a vector containing these values. The default is "status"

interval

a vector of length two specifying the lower and upper quantiles of the interval. The default is c(-Inf, Inf).

age_range

a vector of ages over which curves should be calculated. Default = 40:100

scale

logical. if TRUE centers and scales the PHS scores to unit variance. Default = FALSE.

inverse

logical. if TRUE calculates the inverse (x * -1) the PHS scores to reverse the direction of effect. Default = FALSE.

Value

A data.frame containing ages, the K-M curve, and the upper and lower confidence intervals

Examples

km_curve <- km_curve(test_data)
if (FALSE) { # \dontrun{
ggplot(km_curve, aes(x = time,
                     y = estimate,
                     ymin = conf.low,
                     ymax = conf.high)) +
                     geom_ribbon() +
                     geom_step()
} # }